+++ config/config_manager.go
@@ -123,6 +123,8 @@ func (cm *ConfigManager) setConfigDefaults(v *viper.Viper) {
 	v.SetDefault("web_server.rate_limits.api", 60)
 	v.SetDefault("web_server.rate_limits.introspect", 20)
 	v.SetDefault("web_server.rate_limits.device_code", 10)
+	v.SetDefault("web_server.rate_limits.password", 3)
+	v.SetDefault("web_server.rate_limits.verify", 3)
 
 	// Database configuration
 	v.SetDefault("database.default", "postgres")